Single int64_t)))))

+++ tasks/fibonacci/fibonacci.cpp
@@ -10,7 +10,7 @@ int64_t CalculateNthFibonacci(int64_t n) {
     int64_t current_val = 1;
     int64_t prev_val = 0;
     for (int count = 2; count <= n; ++count) {
-        int next_member = current_val + prev_val;
+        int64_t next_member = current_val + prev_val;
         prev_val = current_val;
         current_val = next_member;
     }
